# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Smart Campos — Módulo G3 (Avaliações e Horários)
  version: 1.0.0
  description: >-
    API autenticada (Bearer) para gestão de avaliações, notas, médias, horários, resultados e impressão de pautas/horários. Consome os serviços de contratos (estudantes, docentes, inscrições e
    financeiro) e bloqueia o lançamento de notas de alunos com dívida.
security:
  - BearerAuth: []
paths:
  /api/v1/auth/login:
    post:
      tags:
        - Autenticação
      summary: Autentica um utilizador e devolve um token de acesso
      description: Público. Verifica credenciais e devolve um accessToken (Bearer, TTL 3600s).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 1
      responses:
        '200':
          description: Credenciais válidas
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      accessToken:
                        type: string
                      tokenType:
                        type: string
                      expiresIn:
                        type: integer
                      user:
                        type: object
                        properties:
                          id:
                            type: string
                            format: uuid
                          role:
                            type: string
                          schoolId:
                            type: string
                            format: uuid
                          name:
                            type: string
                          email:
                            type: string
                  meta:
                    type: object
                    properties:
                      correlationId:
                        type: string
                      page:
                        type: number
                      pageSize:
                        type: number
                      total:
                        type: number
                required:
                  - data
                  - meta
        '400':
          description: Requisição inválida
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
                  details:
                    type: array
                    items:
                      type: object
                  correlationId:
                    type: string
        '401':
          description: Credenciais inválidas ou utilizador inativo
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
                  details:
                    type: array
                    items:
                      type: object
                  correlationId:
                    type: string
      security: []
  /api/v1/results/calculation-methods:
    get:
      tags:
        - Resultados
      summary: Lista os métodos de cálculo disponíveis
      description: Retorna os 6 métodos de cálculo suportados pelo motor académico.
      responses:
        '200':
          description: Operação realizada com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        code:
                          type: string
                          enum:
                            - ARITHMETIC_MEAN
                            - WEIGHTED_PERCENTAGE
                            - PERCENTAGE_SUM
                            - NORMALIZED_WEIGHTED_MEAN
                            - COMPONENT_BASED
                            - CUSTOM_WEIGHTED
                        name:
                          type: string
                        description:
                          type: string
                        formula:
                          type: string
                  meta:
                    type: object
                    properties:
                      correlationId:
                        type: string
                      page:
                        type: number
                      pageSize:
                        type: number
                      total:
                        type: number
                required:
                  - data
                  - meta
        '201':
          description: Recurso criado
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        code:
                          type: string
                          enum:
                            - ARITHMETIC_MEAN
                            - WEIGHTED_PERCENTAGE
                            - PERCENTAGE_SUM
                            - NORMALIZED_WEIGHTED_MEAN
                            - COMPONENT_BASED
                            - CUSTOM_WEIGHTED
                        name:
                          type: string
                        description:
                          type: string
                        formula:
                          type: string
                  meta:
                    type: object
                    properties:
                      correlationId:
                        type: string
                      page:
                        type: number
                      pageSize:
                        type: number
                      total:
                        type: number
                required:
                  - data
                  - meta
        '400':
          description: Requisição inválida (validação Zod ou regra de domínio)
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
                  details:
                    type: array
                    items:
                      type: object
                  correlationId:
                    type: string
        '401':
          description: Não autenticado — token em falta ou inválido (UNAUTHENTICATED)
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
                  details:
                    type: array
                    items:
                      type: object
                  correlationId:
                    type: string
        '404':
          description: Recurso não encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
                  details:
                    type: array
                    items:
                      type: object
                  correlationId:
                    type: string
        '409':
          description: Conflito (duplicação, horário, bloqueio por notas)
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
                  details:
                    type: array
                    items:
                      type: object
                  correlationId:
                    type: string
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
                  details:
                    type: array
                    items:
                      type: object
                  correlationId:
                    type: string
...
  /api/v1/me/financial-status:
    get:
      tags:
        - Resultados
      summary: Estado financeiro do próprio utilizador (estudante)
      description: 'Usado pelo frontend para refletir bloqueio de consulta de notas sem expor montantes: devolve apenas ACTIVE ou BLOCKED.'
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Operação realizada com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      status:
                        type: string
                        enum:
                          - ACTIVE
                          - BLOCKED
                      checkedAt:
                        type: string
                        format: date-time
                    required:
                      - status
                      - checkedAt
                  meta:
                    type: object
                    properties:
                      correlationId:
                        type: string
                      page:
                        type: number
                      pageSize:
                        type: number
                      total:
                        type: number
                required:
                  - data
                  - meta
        '201':
          description: Recurso criado
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      status:
                        type: string
                        enum:
                          - ACTIVE
                          - BLOCKED
                      checkedAt:
                        type: string
                        format: date-time
                    required:
                      - status
                      - checkedAt
                  meta:
                    type: object
                    properties:
                      correlationId:
                        type: string
                      page:
                        type: number
                      pageSize:
                        type: number
                      total:
                        type: number
                required:
                  - data
                  - meta
        '400':
          description: Requisição inválida (validação Zod ou regra de domínio)
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
                  details:
                    type: array
                    items:
                      type: object
                  correlationId:
                    type: string
        '401':
          description: Não autenticado — token em falta ou inválido (UNAUTHENTICATED)
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
                  details:
                    type: array
                    items:
                      type: object
                  correlationId:
                    type: string
        '403':
          description: Estudante com dívida financeira — bloqueado de consultar notas (FINANCIAL_ACCESS_BLOCKED)
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
                  details:
                    type: array
                    items:
                      type: object
                  correlationId:
                    type: string
        '404':
          description: Recurso não encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
                  details:
                    type: array
                    items:
                      type: object
                  correlationId:
                    type: string
        '409':
          description: Conflito (duplicação, horário, bloqueio por notas)
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
                  details:
                    type: array
                    items:
                      type: object
                  correlationId:
                    type: string
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
                  details:
                    type: array
                    items:
                      type: object
                  correlationId:
                    type: string
        '503':
          description: Serviço financeiro indisponível — falha fechada (FINANCIAL_VERIFICATION_UNAVAILABLE)
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
                  details:
                    type: array
                    items:
                      type: object
                  correlationId:
                    type: string
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT